Versão 1.9 03/12/2021

+++ src/Controller/SQLQuery.sql
@@ -6,14 +6,14 @@ create table tb_user(
 	telefone varchar(100),
 	data_nasc varchar(100),
 	senha varchar (100),
-        saldo double,
+    saldo float,
 	constraint pk_user_cpf primary key (cpf)
 );
 
 create table tb_empresa(
         id serial,
  	nome varchar(255),
- 	valor_atual double,
+ 	valor_atual float,
 	constraint pk_empresa_id primary key (id)
 );
 
@@ -21,7 +21,7 @@ create table tb_propriedade(
         id_empresa int,
         cpf_user varchar(100),
         quantidade int,
-        valor_uni double,
+        valor_uni float,
 	constraint fk_user_cpf foreign key (cpf_user) references tb_user(cpf),
         constraint fk_empresa_id foreign key (id_empresa) references tb_empresa(id)
 );
@@ -30,7 +30,7 @@ create table tb_ofertas (
 	cpf_user varchar(100),
 	id_empresa integer,
 	quantidade integer,
-	valor double,
+	valor float,
 	constraint fk_user_cpf foreign key (cpf_user) references tb_user(cpf),
         constraint fk_empresa_id foreign key (id_empresa) references tb_empresa(id)
 );
